[refactor] Rank 클래스 분리
Board 클래스에서 내부 클래스로 Rank 클래스를 가지고 있었습니다.
요구사항 구현으로 Board 자체의 복잡도가 증가함에 따라 가독성을 해치는 문제가 있었습니다.
따라서 Rank 클래스를 별도의 파일로 분리했습니다.
